requirements.txt
Flask
Flask-SQLAlchemy
google-cloud-storage
firebase-admin
lxml
